+++ src/components/SideBar/SideBar.tsx
@@ -1,7 +1,8 @@
-import { Box, Typography, Link } from "@mui/material";
+import { Box } from "@mui/material";
 import React from "react";
 import SearchBar from "../Common/SearchBar";
 import SuggestedUserItem from "./SuggestedUsers";
+import SidebarFooter from "./SidebarFooter";
 
 const styles = {
   ad: {
@@ -23,63 +24,18 @@ const styles = {
     overflow: "hidden",
     width: "100%",
   },
-  adTitle: {
-    color: "primary.main",
-    fontWeight: "bold",
-    textAlign: "center",
-  },
-  footerContainer: {
-    height: "5%",
-    width: "100%",
-  },
-  footerItem: {
-    color: "gray.main",
-    paddingRight: "0.75rem",
-  },
-  footerItemText: {
-    fontSize: "0.75rem",
-  },
-  footerNav: {
-    display: "flex",
-    flexWrap: "wrap",
-    height: "100%",
-    paddingX: 2,
-  },
-  headerTitle: {
-    fontWeight: "bold",
-    paddingX: 2,
-    paddingTop: 1,
-  },
   rightContent: {
     alignItems: "center",
     display: "flex",
     flexDirection: "column",
     gap: 2,
     height: "100vh",
-    paddingLeft: "40px",
     position: "fixed",
     width: "310px",
   },
   searchBarContainer: {
     height: "5%",
     width: "100%",
-  },
-  suHeader: {
-    width: "100%",
-  },
-  suList: {
-    display: "flex",
-    flexDirection: "column",
-    flex: 1,
-  },
-  suggestedUserContainer: {
-    backgroundColor: "gray.light",
-    borderRadius: 5,
-    display: "flex",
-    flexDirection: "column",
-    height: "38%",
-    width: "100%",
-    overflow: "hidden",
   },
 };
 
@@ -89,89 +45,15 @@ const SideBar = () => {
       <Box sx={styles.searchBarContainer}>
         <SearchBar placeholder="Search Chirp" />
       </Box>
-
-      <Box sx={styles.suggestedUserContainer}>
-        <Box sx={styles.suHeader}>
-          <Typography variant="h6" sx={styles.headerTitle}>
-            Who to follow
-          </Typography>
-        </Box>
-        <Box sx={styles.suList}>
-          <SuggestedUserItem />
-        </Box>
-      </Box>
-
+      <SuggestedUserItem />
       <Box sx={styles.adSpaceContainer}>
         <img
           style={styles.ad}
           src={process.env.PUBLIC_URL + "/Ad Gray.gif"}
-          alt="Logo"
+          alt="Advertisement"
         />
       </Box>
-      <Box sx={styles.footerContainer}>
-        <Box sx={styles.footerNav}>
-          <Link
-            underline="hover"
-            target="_blank"
-            href="/coming-soon"
-            rel="noreferrer"
-            sx={styles.footerItem}
-          >
-            <Typography variant="subtitle2" sx={styles.footerItemText}>
-              Terms of Service
-            </Typography>
-          </Link>
-          <Link underline="hover" target="_blank" sx={styles.footerItem}>
-            <Typography variant="subtitle2" sx={styles.footerItemText}>
-              Privacy Policy
-            </Typography>
-          </Link>
-          <Link
-            underline="hover"
-            target="_blank"
-            href="/coming-soon"
-            rel="noreferrer"
-            sx={styles.footerItem}
-          >
-            <Typography variant="subtitle2" sx={styles.footerItemText}>
-              Contact Us
-            </Typography>
-          </Link>
-          <Link
-            underline="hover"
-            target="_blank"
-            href="/coming-soon"
-            rel="noreferrer"
-            sx={styles.footerItem}
-          >
-            <Typography variant="subtitle2" sx={styles.footerItemText}>
-              About The Team
-            </Typography>
-          </Link>
-          <Link
-            underline="hover"
-            target="_blank"
-            href="https://github.com/orgs/Project-Chirp/repositories"
-            rel="noreferrer"
-            sx={styles.footerItem}
-          >
-            <Typography variant="subtitle2" sx={styles.footerItemText}>
-              GitHub Repo
-            </Typography>
-          </Link>
-          <Link
-            underline="hover"
-            target="_blank"
-            href="/coming-soon"
-            rel="noreferrer"
-            sx={styles.footerItem}
-          >
-            <Typography variant="subtitle2" sx={styles.footerItemText}>
-              Docs
-            </Typography>
-          </Link>
-        </Box>
-      </Box>
+      <SidebarFooter />
     </Box>
   );
 };
