+++ src/components/SideBar/SuggestedUsers.tsx
@@ -2,18 +2,17 @@ import {
   Avatar,
   Button,
   List,
-  ListItem,
+  ListItemButton,
   ListItemAvatar,
   ListItemText,
-  Link,
+  Box,
+  Typography,
 } from "@mui/material";
 import React from "react";
 import { useNavigate } from "react-router-dom";
 
 const styles = {
   avatar: {
-    border: "none",
-    cursor: "pointer",
     opacity: 0.75,
     "&:hover": {
       opacity: 1,
@@ -26,9 +25,12 @@ const styles = {
   },
   listText: {
     color: "black",
+    fontWeight: "bold",
+    ":hover": {
+      textDecoration: "underline",
+    },
   },
   userContainer: {
-    cursor: "pointer",
     height: "25%",
     width: "100%",
     display: "flex",
@@ -39,6 +41,7 @@ const styles = {
   },
   followButton: {
     boxShadow: "none",
+    zIndex: 3,
     "&:hover": {
       boxShadow: "none",
       backgroundColor: "primary.main",
@@ -47,6 +50,24 @@ const styles = {
   displayName: {
     fontWeight: "bold",
   },
+  listContainer: {
+    display: "flex",
+    flexDirection: "column",
+    flex: 1,
+  },
+  suggestedUserContainer: {
+    backgroundColor: "gray.light",
+    borderRadius: 5,
+    display: "flex",
+    flexDirection: "column",
+    height: "38%",
+    width: "100%",
+    overflow: "hidden",
+  },
+  suTitle: { fontWeight: "bold", paddingX: 2, paddingTop: 1 },
+  suHeader: {
+    width: "100%",
+  },
 };
 
 const usersData = [
@@ -54,7 +75,6 @@ const usersData = [
     id: 1,
     displayName: "John Doe",
     username: "john_doe",
-    avatarUrl: "https://example.com/avatar1.jpg",
   },
   {
     id: 2,
@@ -76,45 +96,55 @@ const usersData = [
 const SuggestedUsers = () => {
   const navigate = useNavigate();
 
+  const handlePropagation = (e: React.MouseEvent) => {
+    e.stopPropagation();
+  };
+
   return (
-    <List sx={styles.list}>
-      {usersData.map((user) => (
-        <ListItem key={user.id} sx={styles.userContainer}>
-          <ListItemAvatar>
-            <Avatar
-              component="button"
-              onClick={() => navigate("/coming-soon")}
-              alt={user.displayName}
-              sx={styles.avatar}
-            />
-          </ListItemAvatar>
-          <ListItemText
-            primary={
-              <Link
-                underline="hover"
-                target="_blank"
-                href="/coming-soon"
-                rel="noreferrer"
-                sx={styles.listText}
+    <Box sx={styles.suggestedUserContainer}>
+      <Box sx={styles.suHeader}>
+        <Typography variant="h6" sx={styles.suTitle}>
+          Who to follow
+        </Typography>
+      </Box>
+      <Box sx={styles.listContainer}>
+        <List sx={styles.list}>
+          {usersData.map((user) => (
+            <ListItemButton
+              key={user.id}
+              sx={styles.userContainer}
+              onClick={() =>
+                navigate(
+                  `//${window.location.hostname}:${window.location.port}/profile`
+                )
+              }
+            >
+              <ListItemAvatar>
+                <Avatar alt={user.displayName} sx={styles.avatar} />
+              </ListItemAvatar>
+              <ListItemText
+                primary={
+                  <Typography sx={styles.listText}>
+                    {user.displayName}
+                  </Typography>
+                }
+                secondary={`@${user.username}`}
+                primaryTypographyProps={styles.displayName}
+              />
+              <Button
+                variant="contained"
+                color="primary"
+                size="small"
+                sx={styles.followButton}
+                onClick={handlePropagation}
               >
-                {user.displayName}
-              </Link>
-            }
-            secondary={`@${user.username}`}
-            primaryTypographyProps={styles.displayName}
-          />
-          <Button
-            className="followButton"
-            variant="contained"
-            color="primary"
-            size="small"
-            sx={styles.followButton}
-          >
-            Follow
-          </Button>
-        </ListItem>
-      ))}
-    </List>
+                Follow
+              </Button>
+            </ListItemButton>
+          ))}
+        </List>
+      </Box>
+    </Box>
   );
 };
