+++ src/pages/Layout.tsx
@@ -11,6 +11,7 @@ const styles = {
     boxSizing: "border-box",
     flex: "0 0 350px",
     minWidth: 0,
+    paddingLeft: "40px",
   },
 };
